requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
httpx==0.28.1
python-multipart==0.0.20
paho-mqtt==2.1.0
